+++ quill/include/quill/sinks/StreamSink.h
@@ -6,7 +6,6 @@
 #pragma once
 
 #include "quill/core/Attributes.h"
-#include "quill/core/Common.h"
 #include "quill/core/Filesystem.h"
 #include "quill/core/LogLevel.h"
 #include "quill/core/QuillError.h"